+++ src/games/mo6tw/savesystem.h
@@ -38,8 +38,8 @@ class SaveSystem : public SaveSystemBase {
   uint32_t GetTipStatus(size_t tipId) const override;
   void SetTipStatus(size_t tipId, bool isLocked, bool isUnread,
                     bool isNew) override;
-  void SetLineRead(int scriptId, int lineId) override;
-  bool IsLineRead(int scriptId, int lineId) const override;
+  void SetLineRead(size_t scriptId, size_t lineId) override;
+  bool IsLineRead(size_t scriptId, size_t lineId) const override;
   void GetReadMessagesCount(int* totalMessageCount,
                             int* readMessageCount) const override;
   void GetViewedEVsCount(int* totalEVCount, int* viewedEVCount) const override;
